Move calculatePortColor to PortUIViewModel
