+++ Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/InputNodeRowViewModel.swift
@@ -61,37 +61,3 @@ extension InputNodeRowObserver {
         }
     }
 }
-
-extension InputNodeRowViewModel {
-    
-    @MainActor
-    func calculatePortColor(canvasItemId: CanvasItemId,
-                            hasEdge: Bool,
-                            hasLoop: Bool,
-                            selectedEdges: Set<PortEdgeUI>,
-                            selectedCanvasItems: CanvasItemIdSet,
-                            drawingObserver: EdgeDrawingObserver) -> PortColor {
-                
-        // Note: inputs always ignore actively-drawn or animated (edge-edit-mode) edges etc.
-        let canvasItemIsSelected = selectedCanvasItems.contains(canvasItemId)
-        let isSelected = canvasItemIsSelected ||
-        
-        // Relies on self.connectedCanvasItems
-        self.isConnectedToASelectedCanvasItem(selectedCanvasItems)
-        
-        // Relies on self.portViewData
-        || self.hasSelectedEdge(selectedEdges: selectedEdges)
-        
-        return PortColor(isSelected: isSelected,
-                         hasEdge: hasEdge,
-                         hasLoop: hasLoop)
-    }
-    
-    @MainActor
-    func hasSelectedEdge(selectedEdges: Set<PortEdgeUI>) -> Bool {
-        guard let portViewData = self.portAddress else {
-            return false
-        }
-        return selectedEdges.contains { $0.to == portViewData }
-    }
-}
