+++ Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/InputPortUIViewModel.swift
@@ -7,9 +7,10 @@
 
 import Foundation
 
+
 // May want a common protocol differentiated by `portAddress` type
 @Observable
-final class InputPortUIViewModel: Identifiable, AnyObject {
+final class InputPortUIViewModel: PortUIViewModel {
     
     let id: InputCoordinate // which node id + port id this is for
     
@@ -66,5 +67,39 @@ extension InputNodeRowViewModel {
         } set(newValue) {
             self.portUIViewModel.connectedCanvasItems = newValue
         }
+    }
+}
+
+extension InputPortUIViewModel {
+    
+    @MainActor
+    func calculatePortColor(canvasItemId: CanvasItemId,
+                            hasEdge: Bool,
+                            hasLoop: Bool,
+                            selectedEdges: Set<PortEdgeUI>,
+                            selectedCanvasItems: CanvasItemIdSet,
+                            drawingObserver: EdgeDrawingObserver) -> PortColor {
+                
+        // Note: inputs always ignore actively-drawn or animated (edge-edit-mode) edges etc.
+        let canvasItemIsSelected = selectedCanvasItems.contains(canvasItemId)
+        let isSelected = canvasItemIsSelected ||
+        
+        // Relies on self.connectedCanvasItems
+        self.isConnectedToASelectedCanvasItem(selectedCanvasItems)
+        
+        // Relies on self.portViewData
+        || self.hasSelectedEdge(selectedEdges: selectedEdges)
+        
+        return PortColor(isSelected: isSelected,
+                         hasEdge: hasEdge,
+                         hasLoop: hasLoop)
+    }
+    
+    @MainActor
+    func hasSelectedEdge(selectedEdges: Set<PortEdgeUI>) -> Bool {
+        guard let portViewData = self.portAddress else {
+            return false
+        }
+        return selectedEdges.contains { $0.to == portViewData }
     }
 }
