+++ Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/OutputPortUIViewModel.swift
@@ -9,7 +9,7 @@ import Foundation
 
 // May want a common protocol differentiated by `portAddress` type
 @Observable
-final class OutputPortUIViewModel: Identifiable, AnyObject {
+final class OutputPortUIViewModel: PortUIViewModel {
     
     let id: OutputCoordinate // which node id + port id this is for
     
@@ -72,5 +72,47 @@ extension OutputNodeRowViewModel {
         } set(newValue) {
             self.portUIViewModel.connectedCanvasItems = newValue
         }
+    }
+}
+
+
+extension OutputPortUIViewModel {
+    /// Note: an actively-drawn edge SITS ON TOP OF existing edges. So there is no distinction between port color vs edge color.
+    /// An actively-drawn edge's color is determined only by:
+    /// 1. "Do we have a loop?" (blue vs theme-color) and
+    /// 2. "Do we have an eligible input?" (highlight vs non-highlighted)
+    @MainActor
+    func calculatePortColor(canvasItemId: CanvasItemId,
+                            hasEdge: Bool,
+                            hasLoop: Bool,
+                            selectedEdges: Set<PortEdgeUI>,
+                            selectedCanvasItems: CanvasItemIdSet,
+                            drawingObserver: EdgeDrawingObserver) -> PortColor {
+                
+        if let drawnEdge = drawingObserver.drawingGesture,
+           drawnEdge.output.id == self.id {
+            let hasEligibleInput = drawingObserver.nearestEligibleInput.isDefined
+            return PortColor(isSelected: hasEligibleInput,
+                             hasEdge: hasEligibleInput,
+                             hasLoop: hasLoop)
+        }
+        
+        // Otherwise, common port color logic applies:
+        else {
+            let canvasItemIsSelected = selectedCanvasItems.contains(canvasItemId)
+            let isSelected = canvasItemIsSelected || self.isConnectedToASelectedCanvasItem(selectedCanvasItems) || self.hasSelectedEdge(selectedEdges: selectedEdges)
+            
+            return PortColor(isSelected: isSelected,
+                             hasEdge: hasEdge,
+                             hasLoop: hasLoop)
+        }
+    }
+    
+    @MainActor
+    func hasSelectedEdge(selectedEdges: Set<PortEdgeUI>) -> Bool {
+        guard let portViewData = self.portAddress else {
+            return false
+        }
+        return selectedEdges.contains { $0.from == portViewData }
     }
 }
